Install Bugsnag
This PR includes everything required to get Bugsnag
set up and working in development and production.

The Bugsnag API key is stored in the Rails credentials,
so in addition to the changes in this PR, I've added
the Rails secret key to Heroku so it can access the
keys in the credentials file.