ActiveStorage S3 setup
This commit configures ActiveStorage to use Amazon S3 for file storage.

At this stage, the configuration is ready and working, but we've not yet
added the `has_one_attached` association to any of our models. We'll do that
in a following commit when we introduce an Employer model with a logo.

The eager loading configuration changes to the test.rb file was required
to ensure that the ActiveStorage modules are loaded correctly in the
test environment, without this, specs using modals with attached files
would fail.

The reason for this is that ActiveStorage is not loaded in time or loaded
in the wrong order.

We should investigate this further at a later date.

Rails docs
https://github.com/rails/rails/tree/main/activestorage
https://github.com/rails/rails/blob/main/activestorage/app/models/active_storage/blob.rb
https://github.com/rails/rails/blob/main/activestorage/app/models/active_storage/attachment.rb

Tutorial
https://www.honeybadger.io/blog/rails-app-aws-s3/