Rename job columns
Here we remove the employer column (string) from the job table since we'll now
be associating jobs with employers. We also rename the company_name
column to board to better reflect the data it holds.

In some cases this will be the employers own website, but in others it will
be a third party job board where the job is posted.

We also update the views to reflect this change.