Rename Job company_name to board & remove employer columns
Here we rename the company_name column to better reflect what this column
in the Job model represents, which is the job board where the job is listed.

In some cases this will be the employers own website, in others it will
be a third party job board.

We also remove the employer column from the Job model as we're about to
introduce a new Employer model to replace it.