Doors Open script returns Employer name & logo url
In this commit, we update the attributes returned by the Doors Open script
to include the employer name and logo URL.

Without these attributes, the app/sidekiq/scrape_show.rb errors when we
try to call the SetEmployer service object.

We should have included these changes in the prior PR linked below.

https://github.com/JHarrison89/sidekiq-scraper-tailwind/pull/29