Redesign the job show page
In this commit, we adjust update the layout of the job shoe page to make
it more visually appealing by giving the job title more prominence and
adding the employer logo to the right of the title to balance the layout.

We also move the job status section from the top of the page to under the
job title and employer name and update the job status partial to include
a status for unsaved jobs, which is the default for any job that has not
been saved/rejected.

This means we'll always show a status badge for every job, which helps to
balance our layout and encourages users to click the save/reject buttons.