Add .tool-version file
Used to manage ASDF versions, run asdf install to use the versions in this file.

When we installed ASDF, we added yarn and nodejs plugins.
The yarn change added the .yarn-integrity file to the node_modules directory
which we don't want to include in our repository so we add it to gitignore.