Logo helper returns employer/board logos from AWS S3
In this commit, we refactor the `logo_helper` method to return logos from AWS S3
for employers and boards.

In the case of boards, we've added a temporary solution to fetch board logos from
AWS S3 directly since we don't yet have a board model to store images against like
we do with Employer.

Once we've created a Board model, we can remove the `FetchImage` class from the helper.

This change means we now fetch all images from AWS S3 and we can delete the logo images
stored in the `app/assets/images/logos` directory.