Lock Tailwind gems to Tailwind V3
Since Tailwind 4 has now been released we lock the gem versions to prevent
them from accidentally being updated.

I spent some time trying to get Tailwind 4 working, which should be possible
given how simple the Tailwind setup/configuration is in this project, but
it didnt seem like a sensible investment of time given the additional
features (filtering job list, tagging jobs and CVs with DeepSeek) that I
want to work on so this project is ready to show to recruiters.