Add association between jobs and boards
In this commit, we touch a lot of files to set up the association between
boards and jobs, refactor the logo helper to use the board logo and update
the sidekiq scrape_show file to find and save a broad to the job record.

Ideally, these changes would be split up into smaller commits, but for the
sake of saving time, I've bundled them all together.