Separate dropdown functions from Navbar controller
In this commit, we move the dropdown functions into their own controller.
We do this because we want to use the dropdown functions in other places in
the application outside of the navbar.

In the next commit, we'll use the new dropdown controller to open and close
the filter menu on the account show page.