Add filter on the account show page
This dropdown menu acts as a filter for the list of jobs.
When clicked, it displays a checkbox for each unique Employer for the jobs
in the view.

When the user selects one or more employers and clicks the "Filter" button,
we make a GET request to the accounts#show action and only return jobs with
matching employers.

We also update the stimulus controller to close the dropdown when the user
clicks outside the dropdown menu.