Add Link to clear the filter checkbox selections
This link makes a simple get request to the account_path, which causes
the page to do a full reload thereby clearing the filter and showing all
the jobs.

Note the `data: {action: "click->dropdown#toggleDropDown:stop"},` this
stops the action of clicking on the link from calling/running the Stimulus
controller; without this, there was a small amount of latency when clicking
the link.