Respond to filter requests with turbo streams
In this commit, we update the controller to respond to turbo streams as
we'll as standard HTML requests.

When we make a standard request by loading the /accounts URL or clearing
the filter, we respond with the HTML response.

When we click the filer button, we make a turbo stream request and respond
with the new show.turbo_stream.erb file. This file contains a turbo stream
that adds a filter badge, and another turbo stream that replaces the list
of jobs on the /accounts page with the updated (filtered) list of jobs.

Since the filter button is making a request to the SHOW action, we use the
GET method, but since GET requests don't support turbo streams, we need to
add the turbo_stream: true option to the form_with helper.