Refactor job list partial parameter format
In this commit, we adjust how we're passing parameters into the job list partial
to use the correct format. the old version only worked because the partial
has access to the instance variables of the parent view. This commit changes
the partial to pass in the parameters as locals correctly.