Scripts have sleep time parameter
This will allow us to controll the amount of time the scripts sleep for
when scraping a a page. The current defaults are taking too long and since
we need to backfill the existing job recods with Employers and Boards by
re-running the scripts on evey job, this seems like the best option.